remove unused webpack-related code and config; reorganize server spec route logic
