+++ client/esbuild.config.mjs
@@ -101,6 +101,7 @@ function nodeLibToBrowser() {
 	const replace =
 		ENV == 'test'
 			? {
+					// generated by webpack.tape.config.mjs
 					tape: import.meta.resolve('./test/tape.bundle.js').replace('file://', '')
 			  }
 			: ENV == 'dev'
