+++ client/package.json
@@ -25,15 +25,12 @@
     "#updated-types": "./shared/types/terms/updated-types.ts"
   },
   "scripts": {
-    "predev0": "node ./test/import-specs.js name=? && rm -f ../public/bin/*proteinpaint.js",
-    "dev0": "rm -rf ../public/bin; webpack --env NODE_ENV=development --watch --progress --color",
     "dev": "./dev.sh",
     "pretest": "./test/pretest.js",
     "test": "./test.sh *.spec.js",
     "test:unit": "./test.sh *.unit.spec.*",
     "test:integration": "./test.sh *.integration.spec.*",
     "tsc": "tsc -p tsconfig-x.json",
-    "examples": "node ./test/import-specs.js name=_x_.examples && npm run tape",
     "//TODO": "# replace rollup with esbuild",
     "rollw": "rm -rf dist && rollup -c ./rollup.config.js -w --bundleConfigAsCjs",
     "build": "rm -rf dist && rollup -c ./rollup.config.js",
