+++ client/termsetting/README.md
@@ -8,4 +8,4 @@ The API is [documented here](https://docs.google.com/document/d/13bU1azXD6Jl_1w0
 
 - Using a browser: http://localhost:3000/testrun.html?dir=termsetting
 
-- Command line: `node ./test/import-specs.js name=termsetting && npm run tape`
+- Command line: `./test.sh termsetting.*.spec.*`
